CourseNotFoundException

+++ Application/Features/Courses/Commands/Delete/DeleteCourseCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using Domain.Abstractions;
 using Domain.Abstractions.Repositories;
+using Domain.Common.Exceptions;
 using MediatR;
 
 namespace Application.Features.Courses.Commands.Delete;
@@ -8,7 +9,13 @@ internal sealed class DeleteCourseCommandHandler(ICourseRepository courseReposit
 {
     public async Task<DeleteCourseResponse> Handle(DeleteCourseCommand request, CancellationToken cancellationToken)
     {
+        
+        //throw new BadRequestException("Bad request");
         var course= await courseRepository.GetByIdAsync(request.Id, cancellationToken);
+        if (course is null)
+        {
+            throw new CourseNotFoundException(request.Id);
+        }
         courseRepository.Delete(course);
         unitOfWork.Commmit();
         return new DeleteCourseResponse(course.Id);
